refactor: split lms learning modules
