+++ packages/shared/src/lms/learning/helpers.ts
@@ -1,0 +1,9 @@
+export {
+  getLectureInstructorName,
+  getCourseLectures,
+  isEnrolled,
+  getCompletedLectureCount,
+  getLectureCount,
+  getCourseProgress,
+  createCourseCard,
+} from './course';
